requirements.txt
streamlit>=1.32.0
cerebras-cloud-sdk>=0.1.0 